Вынесение констант , рефакторинг кода

+++ museums/app/src/main/java/com/example/museums/API/models/user/ExistingUser.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ExistingUser extends NewUser {
-    long id;
+    Integer id;
 
 
     public ExistingUser(Integer id, String login, String password) {
